Handle expired access token when reading user config
And attempt to refresh the token

Split off code for refreshing token from create token.

Move code that determine whether to use new or old API to earlier when config is being created. The value is then stored in the config for later use.

Use token ID only when revoking token

More logging with func context